+++ src/ben42code/myitertools/__init__.py
@@ -1,5 +1,6 @@
+from ._collapse import collapse
 from ._islice_extended import islice_extended
 from ._iteratorcounter import IteratorCounter
 from ._streamsequence import StreamSequence
 
-__all__ = ["islice_extended", "IteratorCounter", "StreamSequence"]
+__all__ = ["collapse", "islice_extended", "IteratorCounter", "StreamSequence"]